Extract SDL2HardwareCursor to its own file.
