+++ OpenRA.Platforms.Default/Sdl2GraphicsDevice.cs
@@ -11,7 +11,6 @@
 
 using System;
 using System.Drawing;
-using System.IO;
 using System.Runtime.InteropServices;
 using OpenRA.Graphics;
 using SDL2;
@@ -185,18 +184,18 @@ namespace OpenRA.Platforms.Default
 					data = scaledData;
 				}
 
-				return new SDL2HardwareCursor(size, data, hotspot);
+				return new Sdl2HardwareCursor(size, data, hotspot);
 			}
 			catch (Exception ex)
 			{
-				throw new SDL2HardwareCursorException("Failed to create hardware cursor `{0}` - {1}".F(name, ex.Message), ex);
+				throw new Sdl2HardwareCursorException("Failed to create hardware cursor `{0}` - {1}".F(name, ex.Message), ex);
 			}
 		}
 
 		public void SetHardwareCursor(IHardwareCursor cursor)
 		{
 			VerifyThreadAffinity();
-			var c = cursor as SDL2HardwareCursor;
+			var c = cursor as Sdl2HardwareCursor;
 			if (c == null)
 				SDL.SDL_ShowCursor((int)SDL.SDL_bool.SDL_FALSE);
 			else
@@ -222,69 +221,6 @@ namespace OpenRA.Platforms.Default
 					WindowScale = width * 1f / WindowSize.Width;
 
 					OnWindowScaleChanged(oldScale, WindowScale);
-				}
-			}
-		}
-
-		class SDL2HardwareCursorException : Exception
-		{
-			public SDL2HardwareCursorException(string message) : base(message) { }
-			public SDL2HardwareCursorException(string message, Exception innerException) : base(message, innerException) { }
-		}
-
-		sealed class SDL2HardwareCursor : IHardwareCursor
-		{
-			public IntPtr Cursor { get; private set; }
-			IntPtr surface;
-
-			public SDL2HardwareCursor(Size size, byte[] data, int2 hotspot)
-			{
-				try
-				{
-					surface = SDL.SDL_CreateRGBSurface(0, size.Width, size.Height, 32, 0x00FF0000, 0x0000FF00, 0x000000FF, 0xFF000000);
-					if (surface == IntPtr.Zero)
-						throw new InvalidDataException("Failed to create surface: {0}".F(SDL.SDL_GetError()));
-
-					var sur = (SDL.SDL_Surface)Marshal.PtrToStructure(surface, typeof(SDL.SDL_Surface));
-					Marshal.Copy(data, 0, sur.pixels, data.Length);
-
-					// This call very occasionally fails on Windows, but often works when retried.
-					for (var retries = 0; retries < 3 && Cursor == IntPtr.Zero; retries++)
-						Cursor = SDL.SDL_CreateColorCursor(surface, hotspot.X, hotspot.Y);
-
-					if (Cursor == IntPtr.Zero)
-						throw new SDL2HardwareCursorException("Failed to create cursor: {0}".F(SDL.SDL_GetError()));
-				}
-				catch
-				{
-					Dispose();
-					throw;
-				}
-			}
-
-			~SDL2HardwareCursor()
-			{
-				Game.RunAfterTick(() => Dispose(false));
-			}
-
-			public void Dispose()
-			{
-				Game.RunAfterTick(() => Dispose(true));
-				GC.SuppressFinalize(this);
-			}
-
-			void Dispose(bool disposing)
-			{
-				if (Cursor != IntPtr.Zero)
-				{
-					SDL.SDL_FreeCursor(Cursor);
-					Cursor = IntPtr.Zero;
-				}
-
-				if (surface != IntPtr.Zero)
-				{
-					SDL.SDL_FreeSurface(surface);
-					surface = IntPtr.Zero;
 				}
 			}
 		}
